requirement.txt
attrs==21.4.0
cycler==0.11.0
fonttools==4.34.4
importlib-resources==5.8.0
jsonschema==4.7.2
kiwisolver==1.4.4
matplotlib==3.5.2
networkx==2.8.5
numpy==1.23.1
packaging==21.3
paho-mqtt==1.6.1
pandas==1.4.3
Pillow==9.2.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
scipy==1.8.1
six==1.16.0
zipp==3.8.1
